Unbreak -fms-extensions with GNU libc headers
GNU libc uses '__uptr' as a member name in C mode, conflicting with the
eponymous MSVC pointer modifier keyword.

Detect and mark the token as an identifier when these specific conditions are
met. __uptr will continue to work as a keyword for the remainder of the
translation unit.

Fixes PR17824.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@195710 91177308-0d34-0410-b5e6-96231b3b80d8