Make clang-format work in emacs buffers with unicode characters.
Emacs functions by default use character positions; convert characters
to offsets when handing parameters to clang-format and convert byte
offsets we get from clang-format back to character positions.
Reworked the code a bit so the 0-based to 1-based offset calculations
are done in the same place where we do the multi-byte to offset mapping.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@226445 91177308-0d34-0410-b5e6-96231b3b80d8