Remove bad test.
We currently don't canonicalize paths in the preprocessed files.
But we do when writing to PCH.
This causes a discrepancy on Windows with the test below.
This test fails even on unix if you change the test to use
`%S//preprocess.h`.

I am led to conclude that the invariant that this test was intending to
test has not been upheld for a while (and may never have been).

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@243602 91177308-0d34-0410-b5e6-96231b3b80d8