[VFS] Add an in-memory file system implementation.
This is a simple file system tree of memory buffers that can be filled by a
client. In conjunction with an OverlayFS it can be used to make virtual
files accessible right next to physical files. This can be used as a
replacement for the virtual file handling in FileManager and which I intend
to remove eventually.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@249315 91177308-0d34-0410-b5e6-96231b3b80d8