[Tooling] Remove dead code.
It took me some time to figure out why this is not working as expected:
std:error_code converts to true if there is an error. This means we never
ever took the generated absolute path, which happens to be the right thing
anyways as it properly works with virtual files. Just remove the whole
thing, relative paths are covered by existing tooling tests.

git-svn-id: https://llvm.org/svn/llvm-project/cfe/trunk@249408 91177308-0d34-0410-b5e6-96231b3b80d8